add _plot_connectivity_circle label orientation test
